Add Windows Pipe stream for Draconity
Allows communication with Draconity over a named pipe on Windows. Named pipes
are the preferred method of communication for local Draconity instances. This is
just a Windows implementation - it won't work on Mac.

Also integrate this into the live Draconity tests when running on Windows.
They'll now run over a TCP stream and the Pipe stream.

Note this commit does not add fine-grained unit tests or functional tests for
the pipe. It's only tested implicitly, by the live integration tests. Proper
tests will need to be added later to cover edge cases.